+++ src/views/DestinationDetailView.vue
@@ -11,6 +11,7 @@ import PostCard from '../components/PostCard.vue';
 import BudgetCalculator from '../components/BudgetCalculator.vue';
 import ItineraryBuilder from '../components/ItineraryBuilder.vue';
 import PackingChecklist from '../components/PackingChecklist.vue';
+import TravelMapVisualizer from '../components/TravelMapVisualizer.vue';
 
 const route = useRoute();
 const travelStore = useTravelStore();
@@ -53,6 +54,7 @@ const relatedPosts = computed(() => {
 
         <BudgetCalculator :destination-id="destination.id" class="margin-top-calc" />
         <ItineraryBuilder :destination-id="destination.id" class="margin-top-calc" />
+        <TravelMapVisualizer :highlight-dest-id="destination.id" class="margin-top-calc" />
         <PackingChecklist class="margin-top-calc" />
       </div>